Redux -- Checkbox
Work on Redux dispatch for checkbox. Need to create new object for state, not send
old one, or component will not render.

Changed aggregate function to accept an object as second parameter to describe
location of unstructured data, instead of an array, for clarity and future
modifications if needed.

Moved hide from utilities to components/Control for clearer directory structure